// js/domain/enums/RiskGroupAnswer.ts
export enum RiskGroupAnswer {
  olderThan65Years = 'olderThan65Years',
  diabetes = 'diabetes',
  arterialHypertension = 'arterialHypertension',
  heartDiseases = 'heartDiseases',
  chronicRespiratoryDiseases = 'chronicRespiratoryDiseases',
  cancer = 'cancer',
  otherChronicDisease = 'otherChronicDisease',
  noneOfTheAbove = 'noneOfTheAbove',
}